fixed problem with big arrays
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@304 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCsymt.h
@@ -136,11 +136,11 @@ typedef enum {
 } DECLARATOR_TYPE;
 
 typedef struct declarator {
-    DECLARATOR_TYPE dcl_type;     /* POINTER,ARRAY or FUNCTION  */
-    short    num_elem;     /* # of elems if type==array  */
-    short    ptr_const :1;   /* pointer is constant        */
-    short    ptr_volatile:1; /* pointer is volatile        */
-    struct link *tspec;     /* pointer type specifier      */
+    DECLARATOR_TYPE dcl_type;       /* POINTER,ARRAY or FUNCTION  */
+    unsigned int    num_elem;       /* # of elems if type==array  */
+    short           ptr_const :1;   /* pointer is constant        */
+    short           ptr_volatile:1; /* pointer is volatile        */
+    struct link     *tspec;         /* pointer type specifier      */
 } declarator ;
 
 #define DECLARATOR   0
